requirements.txt
astroid==2.11.7
attrs==25.3.0
bandit==1.8.3
beartype==0.20.2
blinker==1.9.0
boolean.py==4.0
bump2version==1.0.1
CacheControl==0.14.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
coverage==7.7.1
cyclonedx-python-lib==8.9.0
defusedxml==0.7.1
dill==0.3.9
docker==7.1.0
execnet==2.1.1
filelock==3.18.0
Flask==3.1.0
Flask-Admin==1.6.1
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
gevent==24.11.1
greenlet==3.1.1
gunicorn==20.1.0
idna==3.10
iniconfig==2.1.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.6
lazy-object-proxy==1.10.0
license-expression==30.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.1.0
natsort==8.4.0
packageurl-python==0.16.0
packaging==24.2
pbr==6.1.1
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.8.0
platformdirs==4.3.7
pluggy==1.5.0
ply==3.11
py==1.11.0
py-serializable==1.1.2
Pygments==2.19.1
pykg-config==1.3.0
pylint==2.13.9
pyparsing==3.2.3
pytest==8.3.5
pytest-cov==6.0.0
pytest-xdist==3.6.1
python-debian==1.0.1
python-dotenv==0.20.0
PyYAML==6.0.2
rdflib==7.1.3
requests==2.32.3
rich==13.9.4
rpm==0.3.1
rpmfile==2.1.0
semantic-version==2.10.0
setuptools==78.1.0
sortedcontainers==2.4.0
spdx-tools==0.8.3
SQLAlchemy==2.0.39
SQLAlchemy-Utils==0.41.2
stevedore==5.4.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
uritools==4.0.3
urllib3==2.3.0
validators==0.34.0
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
xmltodict==0.14.2
zope.event==5.0
zope.interface==7.2
